+++ jsdoc/tables.js
@@ -4,7 +4,7 @@
 /**
 @typedef tableDefinition
 @description Specifies options for table access
-@property {boolean} authorise=false - Execute the {@link module:azure-mobile-apps/express/middleware/authorise authorise middleware} for each table operation
+@property {boolean} authorize=false - Execute the {@link module:azure-mobile-apps/express/middleware/authorize authorize middleware} for each table operation
 @property {boolean} autoIncrement=false - Automatically increment the id column on each insert
 @property {boolean} dynamicSchema=true - Dynamically create table schema
 @property {string} name - Name of the table
