+++ test/E2E/index.js
@@ -14,7 +14,7 @@ var path = require('path'),
 mobileApp = mobileApps(config);
 
 // tables
-mobileApp.tables.add('authenticated', { authorise: true });
+mobileApp.tables.add('authenticated', { authorize: true });
 mobileApp.tables.add('blog_comments', { columns: { postId: 'string', commentText: 'string', name: 'string', test: 'number' } });
 mobileApp.tables.add('blog_posts', { columns: { title: 'string', commentCount: 'number', showComments: 'boolean', data: 'string' } });
 mobileApp.tables.add('dates', { columns: { date: 'date', dateOffset: 'date' } });
@@ -22,7 +22,7 @@ mobileApp.tables.add('movies', { columns: { title: 'string', duration: 'number',
 mobileApp.tables.add('IntIdRoundTripTable', { autoIncrement: true, columns: { name: 'string', date1: 'date', bool: 'boolean', integer: 'number', number: 'number' } });
 mobileApp.tables.add('intIdMovies', { autoIncrement: true, columns: { title: 'string', duration: 'number', mpaaRating: 'string', releaseDate: 'date', bestPictureWinner: 'boolean' } });
 mobileApp.tables.add('OfflineReady');
-mobileApp.tables.add('OfflineReadyNoVersionAuthenticated', { authorise: true });
+mobileApp.tables.add('OfflineReadyNoVersionAuthenticated', { authorize: true });
 mobileApp.tables.import('tables');
 
 app.use(mobileApp);
